change int to float in 835 posting detail page

+++ library/edihistory/edih_835_html.php
@@ -26,7 +26,7 @@
  */
 function edih_round_cb(&$v, $k)
 {
-    $v = round((int)$v, 2);
+    $v = round((float)$v, 2);
 }
 /**
  * Create summary html string for an x12 835 claim payment
